% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliot.r
\name{elliot_wave}
\alias{elliot_wave}
\title{Elliot wave - puglisij}
\usage{
elliot_wave(v, eps = 1e-16, maxiter = 10000, fib = c(0.236, 0.382, 0.5,
  0.618), time_restrictive = TRUE)
}
\arguments{
\item{v}{vector of instrument price data}

\item{eps}{numeric}

\item{maxiter}{maximum number of iterations}

\item{fib}{vector of fibonacci ratios}

\item{time_restrictive}{boolean each wave is restricted to its individual time interval}
}
\value{
list containing entries and golden ratio computed position scaled
}
\description{
Criteria must match as follows: Climb in instrument begins as wave 1, followed by a retracement
which is wave 2, followed by another climb wave 3, followed by a retracement wave 4, and
then followed by the final climb, wave 5.
}
\author{
puglisij
}

